Separate utility functions from main
